Move catalog of test objects into testing package
